+++ src/main/java/com/project/cheerha/common/scheduler/repository/TaskRepository.java
@@ -1,17 +1,8 @@
 package com.project.cheerha.common.scheduler.repository;
 
 import java.time.Instant;
-import java.util.concurrent.TimeUnit;
 
 public interface TaskRepository {
-
-    //Key-Value
-    String getValue(String key);
-    void setValue(String key, String value);
-
-    //Lock
-    boolean tryLock(String lockKey, long waitTime, long leaseTime, TimeUnit unit) throws InterruptedException;
-    void unlock(String lockKey);
 
     //Scheduler
     void saveLastScheduledTime(String key, long time);
